requirements.txt
streamlit 
Pillow
io
base64
numpy 
torch 
torchvision
platform 
os 
